Added protobuffer spec.
Probably not right. Couldn't find a reference to the actual
'basic IPFS object' .proto file anywhere in the IPFS repos or
the libp2p repos, so I wasn't sure what the format was of the
original object.

Also updated the README object examples to match the changed names
for consistency with the rest of the IPFS project.